% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_df.R
\name{ar_df}
\alias{ar_df}
\title{Create AR(p) model data.frame}
\usage{
ar_df(data, y = NULL, p, sort = NULL, formula = FALSE, na.rm = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} or something coercible to one.}

\item{y}{If \code{data} is a \code{data.frame} then this is the character name of the column of the output variable. This does not need to be specified for other types of input.}

\item{p}{The number of lags.}

\item{sort}{If \code{data} is a \code{data.frame} then this is the character name of the column by which to order the data using \code{\link[dplyr:arrange]{dplyr::arrange()}}.
This does not need to be specified for other types of input.}

\item{formula}{A boolean:
\itemize{
\item \code{TRUE}: In addition to the \code{data.frame}, a formula object will be returned with containing the specified model formula.
\item \code{FALSE} (the default): Just \code{data.frame} returned.
}}

\item{na.rm}{A boolean:
\itemize{
\item \code{TRUE}: Removes missing values from the \code{data.frame} using \code{\link[tidyr:drop_na]{tidyr::drop_na()}}.
\item \code{FALSE} (the default): Missing values are returned in \code{data.frame}.
}}
}
\value{
A \code{data.frame} or \code{list}.
}
\description{
\code{ar_df()} Creates new columns from a data-frame or creates a data-frame which contain the predictor variables for an Autoregressive model of a given order.
In addition a formula object can be returned which can be used in standard fit functions such as: \code{\link[stats:lm]{stats::lm()}}, \code{\link[lme4:glmer]{lme4::glmer()}} and \code{\link[brms:brm]{brms::brm()}}.
}
